refactor: type 클래스 위치 변경
