+++ src/main/java/com/example/Dingle/user/dto/CustomUserDetails.java
@@ -1,6 +1,6 @@
 package com.example.Dingle.user.dto;
 
-import com.example.Dingle.user.entity.UserEntity;
+import com.example.Dingle.user.entity.User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -9,10 +9,10 @@ import java.util.Collection;
 
 public class CustomUserDetails implements UserDetails {
 
-    private final UserEntity userEntity;
+    private final User user;
 
-    public CustomUserDetails(UserEntity userEntity) {
-        this.userEntity = userEntity;
+    public CustomUserDetails(User user) {
+        this.user = user;
     }
 
     @Override
@@ -23,7 +23,7 @@ public class CustomUserDetails implements UserDetails {
         collection.add(new GrantedAuthority() {
             @Override
             public String getAuthority() {
-                return userEntity.getUserRole().toString();
+                return user.getUserRole().toString();
             }
         });
         return collection;
@@ -31,12 +31,12 @@ public class CustomUserDetails implements UserDetails {
 
     @Override
     public String getPassword() {
-        return userEntity.getPassword();
+        return user.getPassword();
     }
 
     @Override
     public String getUsername() {
-        return userEntity.getUserId();
+        return user.getUserId();
     }
 
     @Override
