--- src/main/java/com/example/Dingle/user/repository/RealtorRepository.java
@@ -1,9 +1,0 @@
-package com.example.Dingle.user.repository;
-
-import com.example.Dingle.user.entity.RealtorEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface RealtorRepository extends JpaRepository<RealtorEntity, Long> {
-}
