+++ src/main/java/com/example/Dingle/user/repository/UserRepository.java
@@ -1,11 +1,11 @@
 package com.example.Dingle.user.repository;
 
-import com.example.Dingle.user.entity.UserEntity;
+import com.example.Dingle.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<User, Long> {
     boolean existsByUserId(String userId);
-    UserEntity findByUserId(String userId);
+    User findByUserId(String userId);
 }
