+++ src/main/java/com/example/Dingle/user/service/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.example.Dingle.user.service;
 
 import com.example.Dingle.user.dto.CustomUserDetails;
-import com.example.Dingle.user.entity.UserEntity;
+import com.example.Dingle.user.entity.User;
 import com.example.Dingle.user.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -19,7 +19,7 @@ public class CustomUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
-        UserEntity userData = userRepository.findByUserId(userId);
+        User userData = userRepository.findByUserId(userId);
 
         if (userData == null) {
             throw new UsernameNotFoundException(userId);
