+++ src/main/java/com/example/Dingle/user/type/PreferredType.java
@@ -1,4 +1,4 @@
-package com.example.Dingle.user.entity;
+package com.example.Dingle.user.type;
 
 public enum PreferredType {
     ONE_ROOM,
